打印和考试记录功能添加

+++ zq-exam-domain/src/main/java/cn/exam/domain/zj/ZjPaperInfo.java
@@ -30,37 +30,37 @@ public class ZjPaperInfo implements Serializable {
     @Column(name ="paper_name")
     private String paperName;
     /**
-     *
+     * 科目
      */
     @Column(name ="subject_id")
     private Integer subjectId;
     /**
-     *
+     * 难度
      */
     @Column(name ="difficulty")
-    private Integer difficulty;
+    private Double difficulty;
     /**
-     *
+     * 班级
      */
     @Column(name ="class_id")
     private Integer classId;
     /**
-     *
+     * 题型
      */
     @Column(name ="sub_type")
     private String subType;
     /**
-     *
+     * 考试日期
      */
     @Column(name ="paper_date")
     private String paperDate;
     /**
-     *
+     * 考试时间
      */
     @Column(name ="exam_date")
     private String examDate;
     /**
-     *
+     * 考试题目
      */
     @Column(name ="paper_num")
     private Integer paperNum;
@@ -84,6 +84,9 @@ public class ZjPaperInfo implements Serializable {
      */
     @Column(name ="update_time")
     private String updateTime;
+    /**
+     * 总分
+     */
     @Column(name ="paper_score")
     private Integer paperScore;
 }
